backend/pyproject.toml
[tool.poetry]
name = "solanai-backend"
version = "0.1.0"
description = ""
authors = ["Walter Reymundo <walter.reymundo@cloudx.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.12"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
fastapi-cors = "^0.0.6"
python-dotenv = "^1.0.0"
apscheduler = "^3.10.4"
requests = "^2.31.0"
langchain = "^0.0.284"
openai = "^0.28.0"
faiss-cpu = "^1.7.4"
tiktoken = "^0.4.0"
pydantic = "^2.3.0"
webdriver-manager = "^4.0.0"
selenium = "^4.12.0"
tweepy = "^4.14.0"
llama-index = "^0.8.22"
llama-hub = "^0.0.27"
spacy = "^3.6.1"
beautifulsoup4 = "^4.12.2"
psycopg = {extras = ["binary"], version = "^3.1.10"}
cachetools = "^5.3.1"
pypdf = "^3.16.0"
transformers = "^4.33.2"
tensorflow-cpu = "^2.13.0"
celery = {extras = ["redis"], version = "^5.3.4"}
numpy = "^1.26.0"
statsmodels = "^0.14.0"
pandas = "^2.1.1"
prophet = "^1.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "uvicorn main:app --reload"

